Move and rename files.
